+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [16.6.1] - 2026-04-14
+
+### Fixed
+
+- use @Async instead of CompletableFuture.runAsync to prevent classloading issues
+- flush to ensure all sends have actually reached the broker and then commit sync
+
 ## [16.6.0] - 2026-04-13
 
 ### Changed